refactor(web): extract shared port-utils (findFreePort + listenWithFallback)
The "ask the kernel for a free port" pattern (bind to port 0, read
.address().port, close) was duplicated as `getRandomPort` /
`findFreePort` across ~14 test files. The new `listenWithFallback`
helper added for the dev-server boot path uses a related-but-distinct
"scan and bind with retry" pattern.

Pull both into `apps/web/src/lib/port-utils.ts`:

  - `findFreePort()`     — kernel-picked free port; for tests.
  - `listenWithFallback(server, startPort, attempts=20)` — scans
                           upward from `startPort`, retrying on
                           `EADDRINUSE`. Robust against TOCTOU.

`start-server.ts` imports `listenWithFallback` from the new module
instead of defining it inline. `tests/dev-server.test.ts` (the only
test added by this PR) imports `findFreePort` instead of carrying its
own `getRandomPort` copy.

The other ~13 test files still carry their own `getRandomPort` copies
— migrating those is independently useful but orthogonal to #477 and
left as a follow-up.
